+++ strahlenwerk_runtime/src/scalar.h
@@ -1,7 +1,7 @@
 #ifndef MATH_SCALAR_H
 #define MATH_SCALAR_H
 
-double distance(double const& s1, double const& s2);
-double mix(double const& s1, double const& s2, double const& t);
+float distance(float const& s1, float const& s2);
+float mix(float const& s1, float const& s2, float const& t);
 
 #endif // MATH_SCALAR_H
