+++ strahlenwerk_runtime/src/vec2.h
@@ -2,20 +2,20 @@
 #define MATH_VEC2_H
 
 struct vec2 {
-	double x, y;
+	float x, y;
 
 	vec2();
-	vec2(double x_, double y_);
+	vec2(float x_, float y_);
 };
 
 vec2 operator+(vec2 const& v1, vec2 const& v2);
 vec2 operator-(vec2 const& v1, vec2 const& v2);
 
-vec2 operator*(double const& s, vec2 const& v);
-vec2 operator*(vec2 const& v, double const& s);
+vec2 operator*(float const& s, vec2 const& v);
+vec2 operator*(vec2 const& v, float const& s);
 
-double length(vec2 const& v);
-double distance(vec2 const& v1, vec2 const& v2);
-vec2 mix(vec2 const& v1, vec2 const& v2, const double& t);
+float length(vec2 const& v);
+float distance(vec2 const& v1, vec2 const& v2);
+vec2 mix(vec2 const& v1, vec2 const& v2, const float& t);
 
 #endif // MATH_VEC2_H
