+++ strahlenwerk_runtime/src/vec3.h
@@ -2,20 +2,20 @@
 #define MATH_VEC3_H
 
 struct vec3 {
-	double x, y, z;
+	float x, y, z;
 
 	vec3();
-	vec3(double x_, double y_, double z_);
+	vec3(float x_, float y_, float z_);
 };
 
 vec3 operator+(vec3 const& v1, vec3 const& v2);
 vec3 operator-(vec3 const& v1, vec3 const& v2);
 
-vec3 operator*(double const& s, vec3 const& v);
-vec3 operator*(vec3 const& v, double const& s);
+vec3 operator*(float const& s, vec3 const& v);
+vec3 operator*(vec3 const& v, float const& s);
 
-double length(vec3 const& v);
-double distance(vec3 const& v1, vec3 const& v2);
-vec3 mix(vec3 const& v1, vec3 const& v2, const double& t);
+float length(vec3 const& v);
+float distance(vec3 const& v1, vec3 const& v2);
+vec3 mix(vec3 const& v1, vec3 const& v2, const float& t);
 
 #endif // MATH_VEC3_H
